EDI
Former-commit-id: 2d1a8c2ab4a6792a33d5880d8ea7201cfc5b6e76

+++ htdocs/bimpcommercial/objects/ObjectLine.class.php
@@ -4605,7 +4605,7 @@ class ObjectLine extends BimpObject
         }
 
         if (BimpTools::isSubmit('line_pa_ht')) {
-            $line_pa = BimpTools::getValue('line_pa_ht');
+            $line_pa = (float) BimpTools::getValue('line_pa_ht');
         } elseif (BimpTools::isSubmit('line_id_fourn_price')) {
             $fournPrice = BimpCache::getBimpObjectInstance('bimpcore', 'Bimp_ProductFournisseurPrice', (int) BimpTools::getValue('line_id_fourn_price'));
             if (BimpObject::objectLoaded($fournPrice)) {
